requirements.txt
django-mptt>=0.10
django_compressor>=2.4
djangorestframework>=3.11
easy-thumbnails>=2.7
